+++ src/elements/link/block-link-static/block-link-static.snapshot.spec.ts
@@ -10,28 +10,30 @@ import '../../icon.js';
 describe(`sbb-block-link-static`, () => {
   let element: SbbBlockLinkStaticElement;
 
-  beforeEach(async () => {
-    element = await fixture(
-      html` <sbb-block-link-static icon-placement="end" size="m">
-        <sbb-icon
-          aria-hidden="true"
-          data-namespace="default"
-          name="chevron-small-right-small"
-          role="img"
-          slot="icon"
-        ></sbb-icon>
-        Travelcards &amp; tickets.
-      </sbb-block-link-static>`,
-    );
-  });
+  describe('renders', () => {
+    beforeEach(async () => {
+      element = await fixture(
+        html` <sbb-block-link-static icon-placement="end" size="m">
+          <sbb-icon
+            aria-hidden="true"
+            data-namespace="default"
+            name="chevron-small-right-small"
+            role="img"
+            slot="icon"
+          ></sbb-icon>
+          Travelcards &amp; tickets.
+        </sbb-block-link-static>`,
+      );
+    });
 
-  it('renders - DOM', async () => {
-    await expect(element).dom.to.be.equalSnapshot();
-  });
+    it('DOM', async () => {
+      await expect(element).dom.to.be.equalSnapshot();
+    });
 
-  it('renders - Shadow DOM', async () => {
-    await expect(element).shadowDom.to.be.equalSnapshot();
-  });
+    it('Shadow DOM', async () => {
+      await expect(element).shadowDom.to.be.equalSnapshot();
+    });
 
-  testA11yTreeSnapshot();
+    testA11yTreeSnapshot();
+  });
 });